taking photo takes you back to the results page where you are then prompted to post on instagram
