+++ app/src/main/java/com/example/mariokarttournamentorganizer/ViewACTActivity.java
@@ -126,6 +126,7 @@ public class ViewACTActivity extends AppCompatActivity {
 
         Intent enterResults = new Intent(this, EnterResultsActivity.class);
         enterResults.putExtra("actTitle", actTitle);
+        enterResults.putExtra("fromPhoto", false);
         startActivity(enterResults);
     }
